🖋 chore: aws common-module 분리
