+++ kl-common/build.gradle
@@ -7,6 +7,10 @@ dependencies {
     implementation 'io.jsonwebtoken:jjwt-api:0.11.5'
     implementation 'io.jsonwebtoken:jjwt-impl:0.11.5'
     implementation 'io.jsonwebtoken:jjwt-jackson:0.11.5'
-
     implementation project(":kl-domain")
+
+    implementation platform('software.amazon.awssdk:bom:2.20.110')
+    implementation 'software.amazon.awssdk:iam:2.20.26'
+    implementation 'software.amazon.awssdk:sso'
+    implementation 'software.amazon.awssdk:ssooidc'
 }
